+++ apps/ui/src/libs/charts/index.ts
@@ -1,3 +1,4 @@
 
 import * as example_charts from "./samples";
 export const EXAMPLE_CHARTS = Object.values(example_charts);
+export * from './storage';
