+++ apps/ui/src/views/index.ts
@@ -1,3 +1,5 @@
 export { ChatView } from "./ChatView";
 export { FlowChartView } from "./FlowChartView";
 export { SettingsView } from "./SettingsView";
+export { ChartCreationView } from "./ChartCreationView";
+export { LogView } from "./LogView";
